project1/requirement.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
greenlet==2.0.1
multipledispatch==0.6.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
platformdirs==2.5.4
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.43
sqlparse==0.4.3
tzdata==2022.6
virtualenv==20.16.7
